Move auto-yield test to platform spec
